+++ dashboard/src/main/home/app-dashboard/new-app-flow/NewAppFlow.tsx
@@ -22,7 +22,6 @@ import SourceSettings from "./SourceSettings";
 import Services from "./Services";
 import EnvGroupArray, { KeyValueType } from "main/home/cluster-dashboard/env-groups/EnvGroupArray";
 import GithubActionModal from "./GithubActionModal";
-import { ActionConfigType } from "shared/types";
 import Error from "components/porter/Error";
 import { PorterJson, PorterYamlSchema, createFinalPorterYaml } from "./schema";
 import { Service } from "./serviceTypes";
@@ -31,14 +30,6 @@ import Link from "components/porter/Link";
 import { PorterApp } from "../types/porterApp";
 
 type Props = RouteComponentProps & {};
-
-const defaultActionConfig: ActionConfigType = {
-  git_repo: "",
-  image_repo_uri: "",
-  git_branch: "",
-  git_repo_id: 0,
-  kind: "github",
-};
 
 interface FormState {
   applicationName: string;
@@ -84,9 +75,7 @@ type Provider =
   };
 const NewAppFlow: React.FC<Props> = ({ ...props }) => {
   const [porterApp, setPorterApp] = useState<PorterApp>(PorterApp.empty());
-  const [porterYamlPath, setPorterYamlPath] = useState("");
 
-  const [imageUrl, setImageUrl] = useState("");
   const [imageTag, setImageTag] = useState("latest");
   const { currentCluster, currentProject } = useContext(Context);
   const [deploying, setDeploying] = useState<boolean>(false);
@@ -94,12 +83,6 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
   const [currentStep, setCurrentStep] = useState<number>(0);
   const [existingStep, setExistingStep] = useState<number>(0);
   const [formState, setFormState] = useState<FormState>(INITIAL_STATE);
-  const [actionConfig, setActionConfig] = useState<ActionConfigType>(defaultActionConfig);
-  const [buildView, setBuildView] = useState<string>("buildpacks");
-  const [branch, setBranch] = useState("");
-  const [dockerfilePath, setDockerfilePath] = useState("./Dockerfile");
-  const [folderPath, setFolderPath] = useState("./");
-  const [buildConfig, setBuildConfig] = useState({});
   const [porterYaml, setPorterYaml] = useState("");
   const [showGHAModal, setShowGHAModal] = useState<boolean>(false);
   const [showGithubConnectModal, setShowGithubConnectModal] = useState<boolean>(
@@ -113,8 +96,6 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
   const [accessLoading, setAccessLoading] = useState(true);
   const [accessError, setAccessError] = useState(false);
   const [accessData, setAccessData] = useState<GithubAppAccessData>({});
-  const [providers, setProviders] = useState([]);
-  const [currentProvider, setCurrentProvider] = useState(null);
   const [hasProviders, setHasProviders] = useState(true);
 
   const [porterJson, setPorterJson] = useState<PorterJson | undefined>(undefined);
@@ -144,7 +125,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
         "<token>",
         {
           step,
-          stack_name: formState.applicationName,
+          stack_name: porterApp.name,
         },
         {
           cluster_id: currentCluster.id,
@@ -195,7 +176,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
         setDetected({
           detected: true,
           message: `Detected ${Object.keys(porterYamlToJson.apps).length
-            } services from porter.yaml`,
+            } service${Object.keys(porterYamlToJson.apps).length === 1 ? "" : "s"} from porter.yaml`,
         });
       } else {
         setDetected({
@@ -208,30 +189,13 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
       console.log("Error converting porter yaml file to input: " + error);
     }
   };
-  const sortProviders = (providers: Provider[]) => {
-    const githubProviders = providers.filter(
-      (provider) => provider.provider === "github"
-    );
 
-    const gitlabProviders = providers.filter(
-      (provider) => provider.provider === "gitlab"
-    );
-
-    const githubSortedProviders = githubProviders.sort((a, b) => {
-      if (a.provider === "github" && b.provider === "github") {
-        return a.name.localeCompare(b.name);
-      }
-    });
-
-    const gitlabSortedProviders = gitlabProviders.sort((a, b) => {
-      if (a.provider === "gitlab" && b.provider === "gitlab") {
-        return a.instance_url.localeCompare(b.instance_url);
-      }
-    });
-    return [...gitlabSortedProviders, ...githubSortedProviders];
-  };
   useEffect(() => {
     let isSubscribed = true;
+
+    if (currentProject == null) {
+      return;
+    }
 
     api
       .getGitProviders("<token>", {}, { project_id: currentProject?.id })
@@ -245,10 +209,6 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
           setHasProviders(false);
           return;
         }
-
-        const sortedProviders = sortProviders(data);
-        setProviders(sortedProviders);
-        setCurrentProvider(sortedProviders[0]);
       })
       .catch((err) => {
         setHasProviders(false);
@@ -264,7 +224,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
     return regex.test(name);
   };
   const handleAppNameChange = (name: string) => {
-    setFormState({ ...formState, applicationName: name });
+    setPorterApp(PorterApp.setAttribute(porterApp, "name", name));
     if (isAppNameValid(name) && Validators.applicationName(name)) {
       setCurrentStep(Math.max(Math.max(currentStep, 1), existingStep));
     } else {
@@ -280,9 +240,9 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
 
   const shouldHighlightAppNameInput = () => {
     return (
-      formState.applicationName !== "" &&
-      (!isAppNameValid(formState.applicationName) ||
-        formState.applicationName.length > 61)
+      porterApp.name !== "" &&
+      (!isAppNameValid(porterApp.name) ||
+        porterApp.name.length > 61)
     );
   };
 
@@ -304,8 +264,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
         formState.envVariables,
         porterJson,
         // if we are using a heroku buildpack, inject a PORT env variable
-        (buildConfig as any)?.builder != null &&
-        (buildConfig as any)?.builder.includes("heroku")
+        porterApp.builder.includes("heroku")
       );
 
       const yamlString = yaml.dump(finalPorterYaml);
@@ -314,9 +273,9 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
         repository: "",
         tag: "",
       };
-      if (imageUrl && imageTag) {
+      if (porterApp.image_repo_uri && imageTag) {
         imageInfo = {
-          repository: imageUrl,
+          repository: porterApp.image_repo_uri,
           tag: imageTag,
         };
       }
@@ -324,32 +283,28 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
       await api.createPorterApp(
         "<token>",
         {
-          repo_name: actionConfig.git_repo,
-          git_branch: branch,
-          git_repo_id: actionConfig?.git_repo_id,
-          build_context: folderPath,
-          builder:
-            buildView === "buildpacks" ? (buildConfig as any)?.builder : "",
-          buildpacks:
-            buildView === "buildpacks"
-              ? (buildConfig as any)?.buildpacks?.join(",") ?? ""
-              : "",
-          dockerfile: buildView === "docker" ? dockerfilePath : "",
-          image_repo_uri: imageUrl,
+          repo_name: porterApp.repo_name,
+          git_branch: porterApp.git_branch,
+          git_repo_id: porterApp.git_repo_id,
+          build_context: porterApp.build_context,
+          builder: porterApp.dockerfile !== "" ? "" : porterApp.builder,
+          buildpacks: porterApp.dockerfile !== "" ? "" : porterApp.buildpacks.join(","),
+          dockerfile: porterApp.dockerfile,
+          image_repo_uri: porterApp.image_repo_uri,
           porter_yaml: base64Encoded,
           override_release: true,
           image_info: imageInfo,
-          porter_yaml_path: porterYamlPath,
+          porter_yaml_path: porterApp.porter_yaml_path,
         },
         {
           cluster_id: currentCluster.id,
           project_id: currentProject.id,
-          stack_name: formState.applicationName,
+          stack_name: porterApp.name,
         }
       );
 
-      if (!actionConfig?.git_repo) {
-        props.history.push(`/apps/${formState.applicationName}`);
+      if (porterApp.repo_name === "") {
+        props.history.push(`/apps/${porterApp.name}`);
       }
 
       // log analytics event that we successfully deployed
@@ -373,7 +328,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
 
   useEffect(() => {
     setFormState({ ...formState, serviceList: [] });
-  }, [actionConfig, branch]);
+  }, [porterApp.git_branch]);
 
   return (
     <CenterWrapper>
@@ -411,11 +366,11 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
                 <Spacer y={0.5} />
                 <Input
                   placeholder="ex: academic-sophon"
-                  value={formState.applicationName}
+                  value={porterApp.name}
                   width="300px"
                   error={
                     shouldHighlightAppNameInput() &&
-                    (formState.applicationName.length > 30
+                    (porterApp.name.length > 30
                       ? "Maximum 30 characters allowed."
                       : 'Lowercase letters, numbers, and "-" only.')
                   }
@@ -447,47 +402,18 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
                     setFormState({ ...formState, selectedSourceType: type });
                   }}
                 />
-                {/* <SourceSettings
-                  source={formState.selectedSourceType}
-                  imageUrl={imageUrl}
-                  setImageUrl={(x) => {
-                    setImageUrl(x);
-                    setCurrentStep(Math.max(currentStep, 1));
-                  }}
-                  imageTag={imageTag}
-                  setImageTag={setImageTag}
-                  actionConfig={actionConfig}
-                  setActionConfig={setActionConfig}
-                  branch={branch}
-                  setBranch={setBranch}
-                  dockerfilePath={dockerfilePath}
-                  setDockerfilePath={setDockerfilePath}
-                  folderPath={folderPath}
-                  setFolderPath={setFolderPath}
-                  setBuildConfig={setBuildConfig}
-                  porterYaml={porterYaml}
-                  setPorterYaml={(newYaml: string) => {
-                    validatePorterYaml(newYaml);
-                  }}
-                  buildView={buildView}
-                  setBuildView={setBuildView}
-                  setCurrentStep={setCurrentStep}
-                  currentStep={currentStep}
-                  porterYamlPath={porterYamlPath}
-                  setPorterYamlPath={setPorterYamlPath}
-                /> */}
                 <SourceSettings
                   source={formState.selectedSourceType}
-                  buildView={buildView}
-                  setBuildView={setBuildView}
                   setPorterYaml={(newYaml: string) => {
                     validateAndSetPorterYaml(newYaml);
                   }}
                   porterApp={porterApp}
                   setPorterApp={setPorterApp}
-                  // these are separate because the porter app takes in the full repo uri which we create from the url+tag later
-                  imageUrl={imageUrl}
-                  setImageUrl={setImageUrl}
+                  imageUrl={porterApp.image_repo_uri}
+                  setImageUrl={(url: string) => {
+                    setPorterApp(PorterApp.setAttribute(porterApp, "image_repo_uri", url));
+                    setCurrentStep(Math.max(currentStep, 2));
+                  }}
                   imageTag={imageTag}
                   setImageTag={setImageTag}
                 />
@@ -562,8 +488,7 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
               </>,
               <Button
                 onClick={() => {
-                  if (imageUrl) {
-                    console.log(porterYaml);
+                  if (porterApp.image_repo_uri) {
                     deployPorterApp();
                   } else {
                     setDeploymentError(undefined);
@@ -587,19 +512,19 @@ const NewAppFlow: React.FC<Props> = ({ ...props }) => {
           <Spacer y={3} />
         </StyledConfigureTemplate>
       </Div>
-      {showGHAModal && (
+      {showGHAModal && currentCluster != null && currentProject != null && (
         <GithubActionModal
           closeModal={() => setShowGHAModal(false)}
-          githubAppInstallationID={actionConfig.git_repo_id}
-          githubRepoOwner={actionConfig.git_repo.split("/")[0]}
-          githubRepoName={actionConfig.git_repo.split("/")[1]}
-          branch={branch}
-          stackName={formState.applicationName}
+          githubAppInstallationID={porterApp.git_repo_id}
+          githubRepoOwner={porterApp.repo_name.split("/")[0]}
+          githubRepoName={porterApp.repo_name.split("/")[1]}
+          branch={porterApp.git_branch}
+          stackName={porterApp.name}
           projectId={currentProject.id}
           clusterId={currentCluster.id}
           deployPorterApp={deployPorterApp}
           deploymentError={deploymentError}
-          porterYamlPath={porterYamlPath}
+          porterYamlPath={porterApp.porter_yaml_path}
         />
       )}
     </CenterWrapper>
@@ -671,75 +596,4 @@ const StyledConfigureTemplate = styled.div`
   height: 100%;
 `;
 
-const ExpandedWrapper = styled.div`
-  margin-top: 10px;
-  width: 100%;
-  border-radius: 3px;
-  border: 1px solid #ffffff44;
-  max-height: 275px;
-`;
-const ListWrapper = styled.div`
-  width: 100%;
-  height: 240px;
-  background: #ffffff11;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 5px;
-  margin-top: 20px;
-  padding: 40px;
-`;
-const A = styled.a`
-  color: #8590ff;
-  text-decoration: underline;
-  margin-left: 5px;
-  cursor: pointer;
-`;
 
-const ConnectToGithubButton = styled.a`
-  width: 180px;
-  justify-content: center;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  font-size: 13px;
-  cursor: pointer;
-  font-family: "Work Sans", sans-serif;
-  color: white;
-  font-weight: 500;
-  padding: 10px;
-  overflow: hidden;
-  white-space: nowrap;
-  margin-top: 25px;
-  border: 1px solid #494b4f;
-  text-overflow: ellipsis;
-  cursor: ${(props: { disabled?: boolean }) =>
-    props.disabled ? "not-allowed" : "pointer"};
-
-  background: ${(props: { disabled?: boolean }) =>
-    props.disabled ? "#aaaabbee" : "#2E3338"};
-  :hover {
-    background: ${(props: { disabled?: boolean }) =>
-    props.disabled ? "" : "#353a3e"};
-  }
-
-  > i {
-    color: white;
-    width: 18px;
-    height: 18px;
-    font-weight: 600;
-    font-size: 12px;
-    border-radius: 20px;
-    display: flex;
-    align-items: center;
-    margin-right: 5px;
-    justify-content: center;
-  }
-`;
-
-const GitHubIcon = styled.img`
-  width: 20px;
-  filter: brightness(150%);
-  margin-right: 10px;
-`;
