+++ dashboard/src/main/home/app-dashboard/new-app-flow/WebTabs.tsx
@@ -1,5 +1,5 @@
 import Input from "components/porter/Input";
-import React, { useEffect, useRef } from "react";
+import React from "react";
 import Text from "components/porter/Text";
 import Spacer from "components/porter/Spacer";
 import TabSelector from "components/TabSelector";
@@ -11,7 +11,6 @@ interface Props {
   service: WebService;
   editService: (service: WebService) => void;
   setHeight: (height: Height) => void;
-  hasFooter?: boolean;
 }
 
 const RESOURCE_HEIGHT_WITHOUT_AUTOSCALING = 373;
@@ -23,7 +22,6 @@ const WebTabs: React.FC<Props> = ({
   service,
   editService,
   setHeight,
-  hasFooter,
 }) => {
   const [currentTab, setCurrentTab] = React.useState<string>("main");
